API: Extract superclass from InclusiveMetricsEvaluator.
